Animate avatar (#13)
* Add animated gradient and adjust button/link styles

Introduced an animated gradient background to AvatarWrapper for a dynamic visual effect. Set fixed widths and centered text for Button and NavSocialResumeLink components to improve layout consistency. Wrapped OpenRolesBadge in a FlexRow for better alignment in the Hero component.

* Update ExpLeft styles with background and border radius

Added a semi-transparent white background and border radius to the ExpLeft component. Also set mobile width to 100% for better responsiveness.

* Slow down AvatarWrapper gradient animation

Increased the duration of the gradientShift animation from 1s to 4s for a smoother visual effect on the AvatarWrapper component.